sqlite_backend/requirements.txt
Django==4.2.23
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-decouple==3.8